Louis tests, features and bugfixes

+++ src/main/java/zenit/ui/tree/TreeContextMenu.java
@@ -4,11 +4,18 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.Node;
 import javafx.scene.control.*;
+import javafx.scene.input.ClipboardContent;
+import javafx.scene.input.Dragboard;
+import javafx.scene.input.TransferMode;
 import main.java.zenit.filesystem.ProjectFile;
 import main.java.zenit.filesystem.helpers.CodeSnippets;
+import main.java.zenit.filesystem.helpers.FileNameHelpers;
 import main.java.zenit.ui.MainController;
 
+import javax.imageio.IIOException;
 import java.io.File;
+import java.io.IOException;
+import java.util.List;
 
 /**
  * Class that extends {@link javafx.scene.control.ContextMenu} with static menu items with dynamic
@@ -29,7 +36,11 @@ public class TreeContextMenu extends ContextMenu implements EventHandler<ActionE
 	private MenuItem deleteItem = new MenuItem("Delete");
 	private MenuItem importJar = new MenuItem("Import jar");
 	private MenuItem properties = new MenuItem("Properties");
-	
+	private MenuItem moveItem = new MenuItem("Move");
+	private MenuItem dropItem = new MenuItem("Drop");
+
+	private FileTreeItem<String> itemToMove = null;
+
 	/**
 	 * Creates a new {@link TreeContextMenu} that can manipulate a specific {@link
 	 * javafx.scene.control.TreeView TreeView} instance and call methods in a specific
@@ -56,12 +67,28 @@ public class TreeContextMenu extends ContextMenu implements EventHandler<ActionE
 		renameItem.setText(renameItemTitle);
 		deleteItem.setText(deleteItemTitle);
 				
-		if (selectedNode.equals("src") && !createItem.getItems().contains(createPackage)) {
+		/*if (!createItem.getItems().contains(createPackage)) {
 			createItem.getItems().add(createPackage);
-		} else {
-			createItem.getItems().remove(createPackage);
-		}
+		}*/
+
 		FileTreeItem<String> selectedItem = (FileTreeItem<String>) treeView.getSelectionModel().getSelectedItem();
+		if (selectedItem.getFile().isDirectory()) {
+			if (!getItems().contains(createItem)) {
+				getItems().add(0, createItem); // Add createItem at the top
+			}
+			if (!createItem.getItems().contains(createPackage)) {
+				createItem.getItems().add(createPackage);
+			}
+			if (!createItem.getItems().contains(createClass)) {
+				createItem.getItems().add(createClass);
+			}
+			if (!createItem.getItems().contains(createInterface)) {
+				createItem.getItems().add(createInterface);
+			}
+		} else {
+			getItems().remove(createItem);
+		}
+
 		if (selectedItem.getType() == FileTreeItem.PROJECT) {
 			getItems().add(importJar);
 			getItems().add(properties);
@@ -94,7 +121,7 @@ public class TreeContextMenu extends ContextMenu implements EventHandler<ActionE
 	private void initContextMenu() {
 		createItem.getItems().add(createClass);
 		createItem.getItems().add(createInterface);
-		getItems().addAll(createItem, renameItem, deleteItem);
+		getItems().addAll(createItem, renameItem, deleteItem, moveItem, dropItem);
 		createClass.setOnAction(this);
 		createInterface.setOnAction(this);
 		renameItem.setOnAction(this);
@@ -102,6 +129,9 @@ public class TreeContextMenu extends ContextMenu implements EventHandler<ActionE
 		createPackage.setOnAction(this);
 		importJar.setOnAction(this);
 		properties.setOnAction(this);
+		moveItem.setOnAction(this);
+		dropItem.setOnAction(this);
+		dropItem.setVisible(false);
 	}
 	
 	/**
@@ -114,9 +144,87 @@ public class TreeContextMenu extends ContextMenu implements EventHandler<ActionE
 				treeView.getSelectionModel().getSelectedItem();
 		File newFile = controller.createFile(parent.getFile(), typeCode);
 		if (newFile != null) {
-			FileTreeItem<String> newItem = new FileTreeItem<String>(newFile, newFile.getName(), FileTreeItem.CLASS);
+			FileTreeItem<String> newItem = new FileTreeItem<String>(newFile, newFile.getName(), FileTreeItem.CLASS, false);
 			parent.getChildren().add(newItem);
+			sortChildren(parent);
 		}
+	}
+
+	private void moveFile(FileTreeItem<String> selectedItem, File destinationDir) {
+		File oldFile = selectedItem.getFile();
+		File newFile = new File(destinationDir, oldFile.getName());
+
+		if (newFile.exists()) {
+			Alert alert = new Alert(Alert.AlertType.ERROR);
+			alert.setTitle("Move File Error");
+			alert.setHeaderText("File Move Failed");
+			alert.setContentText("A file with the name \"" + oldFile.getName() + "\" already exists in the destination directory.");
+			alert.showAndWait();
+			return;
+		}
+
+		newFile = controller.moveFile(oldFile, destinationDir);
+		if (newFile != null) {
+			selectedItem.setFile(newFile);
+			selectedItem.setValue(newFile.getName());
+			FileTree.changeFileForNodes(selectedItem, newFile);
+
+			TreeItem<String> parent = selectedItem.getParent();
+			parent.getChildren().remove(selectedItem);
+
+			FileTreeItem<String> newParent = findTreeItem(destinationDir);
+			if (newParent != null) {
+				newParent.getChildren().add(selectedItem);
+			}
+
+			String newPackageName = FileNameHelpers.getPackagenameFromFile(newFile);
+			try {
+				FileNameHelpers.updatePackageName(newFile, newPackageName);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+
+			updateOpenTabs(oldFile, newFile);
+		}
+	}
+
+	public void updateOpenTabs(File oldFile, File newFile) {
+		TabPane tabPane = controller.getTabPane();
+		Tab oldTab = findTabByFile(tabPane, oldFile);
+		if (oldTab != null) {
+			tabPane.getTabs().remove(oldTab);
+			controller.openFile(newFile);
+		}
+	}
+
+	private Tab findTabByFile(TabPane tabPane, File file) {
+		String filePath = file.getAbsolutePath();
+		for (Tab tab : tabPane.getTabs()) {
+			if (tab.getUserData() != null) {
+				String tabFilePath = tab.getUserData().toString();
+				if (tabFilePath.equals(filePath)) {
+					return tab;
+				}
+			}
+		}
+		return null;
+	}
+
+	private FileTreeItem<String> findTreeItem(File file) {
+		return findTreeItem((FileTreeItem<String>) treeView.getRoot(), file);
+	}
+
+	private FileTreeItem<String> findTreeItem(FileTreeItem<String> root, File file) {
+		if (root.getFile().equals(file)) {
+			return root;
+		}
+		for (TreeItem<String> child : root.getChildren()) {
+			FileTreeItem<String> result = findTreeItem((FileTreeItem<String>) child, file);
+			if (result != null) {
+				return result;
+			}
+		}
+		return null;
 	}
 
 	/**
@@ -138,22 +246,65 @@ public class TreeContextMenu extends ContextMenu implements EventHandler<ActionE
 				selectedItem.setFile(newFile);
 				selectedItem.setValue(newFile.getName());
 				FileTree.changeFileForNodes(selectedItem, selectedItem.getFile());
+				updateOpenTabs(selectedFile, newFile);
 			}
 		} else if (actionEvent.getSource().equals(deleteItem)) {
+			Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+			alert.setTitle("Delete File");
+			alert.setHeaderText("Are you sure you want to delete this file?");
+			alert.setContentText("File: " + selectedFile.getName());
+
+			ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
+			if (result == ButtonType.OK) {
 			controller.deleteFile(selectedFile);
 			selectedItem.getParent().getChildren().remove(selectedItem);
+			}
 		} else if (actionEvent.getSource().equals(createPackage)) {
 			File packageFile = controller.newPackage(selectedFile);
 			if (packageFile != null) {
-				FileTreeItem<String> packageNode = new FileTreeItem<String>(packageFile, packageFile.getName(), FileTreeItem.PACKAGE);
+				FileTreeItem<String> packageNode = new FileTreeItem<String>(packageFile, packageFile.getName(), FileTreeItem.PACKAGE, false);
 				selectedItem.getChildren().add(packageNode);
+				sortChildren(selectedItem);
 			}
 		} else if (actionEvent.getSource().equals(importJar)) {
 			ProjectFile projectFile = new ProjectFile(selectedFile.getPath());
-			controller.chooseAndImportLibraries(projectFile);
+			List<File> libraries = controller.chooseAndImportLibraries(projectFile);
+			if (!libraries.isEmpty()) {
+				FileTreeItem<String> libNode = findTreeItem(new File(selectedFile, "lib"));
+
+				if (libNode != null) {
+					for (File jarFile : libraries) {
+						File newJar = new File(libNode.getFile(), jarFile.getName());
+						FileTreeItem<String> jarNode = new FileTreeItem<>(newJar, newJar.getName(), FileTreeItem.FILE, false);
+						libNode.getChildren().add(jarNode);
+					}
+					sortChildren(libNode);
+				}
+			}
 		} else if (actionEvent.getSource().equals(properties) && selectedItem.getType() == FileTreeItem.PROJECT) {
 			ProjectFile projectFile = new ProjectFile(selectedFile.getPath());
 			controller.showProjectProperties(projectFile);
+		} else if (actionEvent.getSource().equals(moveItem)) {
+			itemToMove = selectedItem;
+			dropItem.setVisible(true);
+		} else if (actionEvent.getSource().equals(dropItem) && itemToMove != null) {
+			moveFile(itemToMove, selectedItem.getFile());
+			itemToMove = null;
+			dropItem.setVisible(false);
 		}
+	}
+
+	private void sortChildren(FileTreeItem<String> parent) {
+		parent.getChildren().sort((o1, o2) -> {
+			FileTreeItem<String> f1 = (FileTreeItem<String>) o1;
+			FileTreeItem<String> f2 = (FileTreeItem<String>) o2;
+			if (f1.getType() == FileTreeItem.PACKAGE && f2.getType() != FileTreeItem.PACKAGE) {
+				return -1;
+			} else if (f1.getType() != FileTreeItem.PACKAGE && f2.getType() == FileTreeItem.PACKAGE) {
+				return 1;
+			} else {
+				return f1.getValue().compareTo(f2.getValue());
+			}
+		});
 	}
 }
